Modifies comment system. Makes comments attachable

+++ db/migrate/20170517062053_drop_categories_table.rb
@@ -1,6 +1,8 @@
 class DropCategoriesTable < ActiveRecord::Migration[5.0]
   def up
-    drop_table :categories
+    drop_table :categories if ActiveRecord::Base
+                              .connection
+                              .data_source_exists? 'categories'
   end
 
   def down
